refactor: renamed react components files to use PascalCase
