+++ client/index.html
@@ -17,6 +17,6 @@
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
 
-    <script type="module" src="/src/index.jsx"></script>
+    <script type="module" src="/src/Index.jsx"></script>
   </body>
 </html>
