+++ client/src/Index.jsx
@@ -4,7 +4,7 @@ import "./assets/index.css";
 import reportWebVitals from "./app/reportWebVitals";
 
 import { RouterProvider } from "react-router-dom";
-import { router } from "./app/router";
+import { router } from "./app/Router";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
